CPT dashboard functional test framework
This adds a mechanism to "can" and restore a small prototype ILAB (Crucible
CDM) Opensearch database in a pod along with the dashboard back end, front
end, and functional tests. The functional tests run entirely within the pod,
with no exposed ports and with unique container and pod names, allowing for
the possibility of simultaneous runs (e.g., a CI) on the same system.

This also has utilities for diagnosing a CDM (v7) datastore and cloning a
limited subset, along with creating an Opensearch snapshot from that data
to bootstrap the functional test pod.

Only a few functional test cases are implemented here, as demonstration. More
will be added separately.
